Keep timestamps in uint32_t, so the host wraps millis() like the ESP32
Every timestamp and interval off millis() was an unsigned long, which is
32 bits on the ESP32 and 64 on the host, so the one arithmetic the tests
could not reach was the wrap at 49.7 days: on the host it never came.
Now they are uint32_t on both, the fake millis() included, and three
tests take the clock through zero: the PI integral gets one more second,
the thermometer low-pass moves by a second's worth, and the 10 s
setpoint refresh comes round rather than waiting out 49 days.

The OpenTherm frame in loop() stays an unsigned long, since that is what
the library returns and no time arithmetic is done on it.

+++ ardu.ino
@@ -31,7 +31,7 @@ bool readCentralHeatingOn = false;
 bool readHotWaterOn = false;
 bool readFlameOn = false;
 OpenThermResponseStatus responseStatus = OpenThermResponseStatus::NONE;
-unsigned long response_ts = 0;
+uint32_t response_ts = 0;
 float readBoilerTemperature = 0.0;
 float readPressure = 0.0;
 float readReturnTemperature = 0.0;
@@ -67,16 +67,16 @@ ChTempSource setChTempSource = CH_TEMP_MANUAL;
 PISource pi;
 
 // state of pushSetpoints(); NAN means nothing was sent yet, so the first pass pushes
-const unsigned long setpointInterval = 10000; // 10 s between refreshes
+const uint32_t setpointInterval = 10000; // 10 s between refreshes
 const float setpointEpsilon = 0.5;  // smallest change worth an early write, so the
                                     // PI output does not write on every pass
-unsigned long lastSetpointSent = 0;
+uint32_t lastSetpointSent = 0;
 float sentBoilerTemperature = NAN;
 float sentDHWTemperature = NAN;
 
 Preferences preferences;
 
-unsigned long lastWifiConnected = 0;
+uint32_t lastWifiConnected = 0;
 int wifi_reconnects = 0;
 
 uint32_t boot_count = 0;
@@ -337,7 +337,7 @@ void loop()
 {
     // check wifi connection and reconnect if needed
     if (WiFi.status() != WL_CONNECTED) {
-        unsigned long now = millis();
+        uint32_t now = millis();
         if (now - lastWifiConnected > 30000) { // try to reconnect every 30 seconds
             Serial.println("WiFi not connected, trying to reconnect...");
             WiFi.disconnect(true);
